bijnafinal

+++ test/project.cs
@@ -45,8 +45,12 @@ namespace test
 
             string getDirectory = Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location);
             string fileName = "BeoordelingData.txt";
-            TextWriter txt = new StreamWriter(getDirectory+"\\"+fileName);
-            txt.Close();
+            if(!File.Exists(getDirectory + "\\" + fileName))
+            {
+                TextWriter txt = new StreamWriter(getDirectory + "\\" + fileName);
+                txt.Close();
+            }
+            
 
             List<string> data = File.ReadAllLines(getDirectory+"\\"+fileName).ToList();
             foreach (string d in data)
